Add merge sort function
This function is an implementation of the pseudocode of the merge sort
algorithm using recursion. It consists in dividing an array until it
has only 1 value, first from the left part and then the right part.
When that happens, it merge both parts in one new array and makes the
sorting process; when that finishes it returns this new sorted array.
It keeps making this three steps until the whole array is sorted.